examples/mindmap-assistant/backend/pyproject.toml
[project]
name = "mindmap-assistant-backend"
version = "0.1.0"
description = "Mindmap Assistant Backend with FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx[socks]>=0.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
